CPM-999: Move classes into ServiceAPI
