+++ components/identifier-generator/back/src/Infrastructure/Symfony/Resources/config/anti_corruption_layers.yml
@@ -1,4 +1,4 @@
 services:
   Akeneo\Pim\Automation\IdentifierGenerator\Infrastructure\AntiCorruptionLayer\CategoriesHaveAtLeastOneChild:
     arguments:
-      - '@Akeneo\Pim\Enrichment\Category\API\Query\CategoriesHaveAtLeastOneChild'
+      - '@Akeneo\Category\ServiceApi\Query\CategoriesHaveAtLeastOneChild'
